+++ kibot/misc.py
@@ -267,6 +267,10 @@ W_BLINDVIAS = '(W118) '
 W_LIBTVERSION = '(W119) '
 W_LIBTUNK = '(W120) '
 W_DRC7BUG = '(W121) '
+W_BADTOL = '(W122) '
+W_BADRES = '(W123) '
+W_RESVALISSUE = '(W124) '
+W_RES3DNAME = '(W125) '
 # Somehow arbitrary, the colors are real, but can be different
 PCB_MAT_COLORS = {'fr1': "937042", 'fr2': "949d70", 'fr3': "adacb4", 'fr4': "332B16", 'fr5': "6cc290"}
 PCB_FINISH_COLORS = {'hal': "8b898c", 'hasl': "8b898c", 'imag': "8b898c", 'enig': "cfb96e", 'enepig': "cfb96e",