+++ tests/test_plot/test_step.py
@@ -155,6 +155,19 @@ def test_render_3d_variant_1(test_dir):
     ctx.clean_up(keep_project=True)
 
 
+# @pytest.mark.slow
+# @pytest.mark.pcbnew
+# def test_render_3d_res_tht_1(test_dir):
+#     prj = 'resistor_tht'
+#     yaml = 'render_3d_tht_res_1'
+#     ctx = context.TestContext(test_dir, prj, yaml)
+#     ctx.run()  # extra_debug=True
+#     # Check all outputs are there
+#     name = prj+'-3D_top.png'
+#     ctx.expect_out_file(name)
+#     ctx.clean_up(keep_project=True)
+
+
 @pytest.mark.slow
 @pytest.mark.pcbnew
 @pytest.mark.skipif(context.ki5(), reason="uses pcb2blender")