+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.6.2] - UNRELEASED
+### Added
+- 3D: colored 3D models for THT resistors
+
 ### Fixed
 - Makefile: don't skip all preflights on each run, just the ones we generate
   as targets. (#405)